backend/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mutagen==1.47.0
mysql-connector-python==8.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.0.1
gunicorn==20.1.0
